refactor(server): #125 — extract trust handler into routes/trust.rs
